requirements.txt
hypothesis>=6.0.0
pytest>=7.0.0
pyyaml>=6.0